FIX: Scroll problem for right-menu
top was relative to the page and when the platform links disappear the relative top changes.
Fix is to make the top fixed at zero and move the content of the nav object down to a set distance (chose 160px as close to original but a little further away from platform links).